Zuul: [WP25EasterEggs] enable recursion
The extension depends on Wikibase and its Selenium tests would fail when
dependencies are missing. Surely we need a better system that would
prevent Wikibase tests from running when sending a patch for
WP25EasterEggs, but it is not possible right now.

Turning on recursion should make the Selenium tests to pass. A drawback
is the PHPUnit tests will be slower since they will run more than
needed.

Bug: T411408
Change-Id: Ie506de70f1f3795ccb4cee446adebc5d3dd90d85