inference-services: add CI pipeline jobs for embeddings model-server
Add CI pipeline jobs to publish the embeddings model-server
image to the Wikimedia docker registry.

'test' and 'gate-and-submit' jobs were not added for this
model-server at this time, as it currently only requires
ci-lint tests, which are executed by the pre-commit-check.
Once unit tests are implemented for this model-server,
we should include the 'test' and 'gate-and-submit' jobs
to enable the execution of ci-unit tests.

Bug: T412338
Change-Id: I608d9bcabb5a98b2ab0ab6a6629ed006d8a72546